Keys

+++ script.js
@@ -45,7 +45,7 @@ class Calculator {
       case '*':
         computation = prev * current;
         break;
-      case '÷':
+      case '/':
         computation = prev / current;
         break;
       default:
@@ -88,7 +88,7 @@ class Calculator {
     }
   }
 }
-
+const allKeys = Array.from(document.querySelectorAll("button"))
 const numberButtons = document.querySelectorAll('[data-number]');
 const operationButtons = document.querySelectorAll('[data-operation]');
 const equalsButton = document.querySelector('[data-equals]');
@@ -100,6 +100,8 @@ const previousOperandTextElement = document.querySelector(
 const currentOperandTextElement = document.querySelector(
   '[data-current-operand]'
 );
+let numbers = ["1","2","3","4","5","6","7","8","9","0"]
+let operators = ["+", "-", "/", "*"];
 
 const calculator = new Calculator(
   previousOperandTextElement,
@@ -135,31 +137,29 @@ deleteButton.addEventListener('click', (button) => {
   calculator.updateDisplay();
 });
 
-numberButtons.forEach((button) => {
-  button.addEventListener('keydown', () => {
-    calculator.appendNumber(button.innerText);
-    calculator.updateDisplay();
-  });
-});
 
-operationButtons.forEach((button) => {
-  button.addEventListener('keydown', () => {
-    calculator.chooseOperation(button.innerText);
-    calculator.updateDisplay();
-  });
-});
-
-equalsButton.addEventListener('keydown', (button) => {
-  calculator.compute();
-  calculator.updateDisplay();
-});
-
-allClearButton.addEventListener('keydown', (button) => {
-  calculator.clear();
-  calculator.updateDisplay();
-});
-
-deleteButton.addEventListener('keydown', (button) => {
-  calculator.delete();
-  calculator.updateDisplay();
-});
+allKeys.forEach(keys=>{
+  keys.addEventListener("keydown", function(event) {
+      let keypress = event.key
+      numbers.forEach(number => {
+          if(keypress == number){
+            calculator.appendNumber(keypress);
+            calculator.updateDisplay();
+          }
+      })   
+      operators.forEach(operator => {
+      if(keypress == operator){
+        calculator.chooseOperation(keypress);
+        calculator.updateDisplay()
+        }
+      })
+      if(keypress == "Escape"){
+        calculator.compute();
+        calculator.updateDisplay();
+      }
+      if(keypress == "Backspace"){
+        calculator.delete();
+        calculator.updateDisplay();
+      }
+  })
+})
